backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
neo4j==5.25.0
python-multipart==0.0.12
python-dotenv==1.0.1
